feat(core): awesome update to support the lower & upper case in changes:-)
Now you can also change word with the upper and the lower case!!!

for example:

now it has a word 'True'
and now you can use the plugin to change it to 'False'
you don't need to add { 'True‘, 'False' } to your dictorary.

so as 'TRUE' to 'FALSE'

It's a great changes.